requirements.txt
Flask==2.2.2
numpy==1.24.1
pandas==1.5.3
python-dotenv==0.21.1
